fix: NftPreview architecture simplified
