+++ components/base/NftCard/index.tsx
@@ -1,2 +1,1 @@
 export { default } from './NftCard';
-export { default as NftCardWithEffects } from './NftCardWithEffects';
