+++ components/base/NftPreview/components/NftUpload.tsx
