+++ components/base/NftPreview/index.tsx
@@ -1,1 +1,3 @@
 export { default } from "./NftPreview";
+export { default as NftCardWithEffects } from './components/NftCardWithEffects';
+export { default as NftUpload } from "./components/NftUpload";
